dbx/pixels/resources/dicom_web_gateway/requirements.txt
fastapi
uvicorn
databricks-sdk
requests
zstd
psycopg2-binary
psutil
